chore: Move to class
